Fix warnings

+++ tutorial/imgproc/count-coins/tutorial-count-coins.cpp
@@ -128,8 +128,8 @@ int main(int argc, char *argv[]) {
       std::stringstream ss;
       ss << "Coin " << nb_coins;
 
-      double centroid_x = std::fabs(obj.get(1, 0) / obj.get(0, 0));
-      double centroid_y = std::fabs(obj.get(0, 1) / obj.get(0, 0));
+      int centroid_x = (int) std::fabs(obj.get(1, 0) / obj.get(0, 0));
+      int centroid_y = (int) std::fabs(obj.get(0, 1) / obj.get(0, 0));
       vpDisplay::displayText(I_draw_contours, centroid_y, centroid_x-20., ss.str(), vpColor::red);
     }
   }
